Add tap actions for ContextElements
